Added some helper functions and adjusted the urdf files.
The pose in ROS now is consistent with the  one in smartPad.
It should be noted that the base and tip link are left_base and left_ee_link ,respectively. The same thing to the right arm.
In smartPad, the Feature should be choosen BASE, and Rotation is RPY[rad].